pyproject.toml
[project]
name = "razgon-drf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "django-cors-headers>=4.3.0",
    "django-htmx>=1.17.0",
]
